u-boot: base on upstream repository
Standard practice in OE has been to base u-boot on the official denx
repository with patches specific to the platform.  Using a gumstix forked
repository and a branch for which the relevant patches are not reflected in
a PR number or any other indicator leads to confusion and the risk of
incompatible packages if that branch is updated.

Revert to using the official repository with a specific base version
(corresponding to tag v2013.07) and specific patches, which should in turn
be pushed upstream so they can be removed from recipes based on later u-boot
releases.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>