u-boot: read boot environment overrides from mmc card
Image-specific environment settings (such as command-line options related to
DSP features) can cause boot problems when mis-applied to images that don't
support them, as happens when the settings are stored on the overo COM
itself.  Back-port a u-boot enhancement (available on most OMAP platforms)
that reads uEnv.txt from the mmc card prior to booting.  This allows
features settings required for certain images used on a particular Gumstix
device to be coupled with the SD card used to boot that image.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>