linux-gumstix 3.5: base on upstream repository
Standard practice in OE layers that do not use linux-yocto has been to base
linux packages on the official stable repository with patches specific to
the platform.

Using a gumstix forked repository and a branch for which the relevant
patches are not reflected in a PR number or any other indicator leads to
confusion and the risk of incompatible packages if that branch is updated.
It is critical for reproducibility and stable package upgrades that the
package version or revision be updated on any content change: processes that
obscure these changes should be avoided.  AUTOREV in particular is a bad
idea.

Revert to using the stable repository with a specific base version with
patches specific to gumstix.  The tag for the base version is expressed in
SRCREV using its SHA identifier.  The patches are maintained in a
non-canonical mirror repository hosted on github.  Addition of new patches
or changing the content of an existing patch without changing PV must be
accompanied by an increment of PR.  PR should be reset to r0 when PV
changes; this is the default, but it is left explicit as a reminder.

The package version now incorporates the base Linux version including third
identifier, although most patches are shared from the linux-gumstix-3.5 as
they should apply to all releases based on the linux-3.5.y branch.

Other than a change in name from linux-gumstix-3.5-r0 to
linux-gumstix-3.5.0-r0 this commit introduces no changes to the kernel or
module contents and the two should be equivalent.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>