linux-gumstix 3.5: backport patch for mt9v032 module problem
The standard kernel build has support for the mt9v032 "Caspa" camera, but if
one is not installed the hardware probe fails.  In this situation a driver
bug caused the module reference count to be decremented too many times,
making it impossible to unload the module, for example to allow the module
supporting a different camera to be loaded instead.

NB: This upstream patch first appeared in Linux 3.9 and was modified to
reflect the location of the sources in the 3.5 kernel.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>