Revert "Using correct xserver component"
This reverts commit f7e8c4c3a5748e34a81cd97ad182c093904651c7.

Couldn't get output to the screen. The xf86-video-omap recipe warns this driver is not ready for prime time:

"NOTE: this driver is work in progress..  you probably don't want to try \
and use it yet.  The API/ABI between driver and kernel, and driver and \
acceleration submodules is not stable yet.  This driver requires the \
omapdrm kernel driver w/ GEM support. \ "