duovero: Use hardfloat tuning
SGX binaries are provided in hardfloat mode only so we need to be
building in hardfloat mode to use these binaries.  Note that we use
"cortexa8hf-neon" rather than "cortexa9hf-neon" as the only difference
is -mtune=cortexa9 which, from these admittedly outdated, results [1],
doesn't have a substantial impact.  This allows all our platforms to
share the same tuning which substantially reduces build time and disk
space when building for Overo, Pepper, and DuoVero.

[1] https://wiki.linaro.org/MichaelHope/Sandbox/CoreMark1

Signed-off-by: Ash Charles <ashcharles@gmail.com>

Conflicts:
	conf/machine/duovero.conf