linux-gumstix: Support LB035q02 display for Palo35
Include the LG.Philips LB035Q02 3.5" LCD in the kernel.  Reduce the
SPI timing for this device in the Palo35 device tree file as it
doesn't work correctly at the theoretical max of 10MHz.  This does not
affect the performance of the screen---just communication with the
screen microcontroller used for initial configuration.

Signed-off-by: Ash Charles <ashcharles@gmail.com>