u-boot: allow modification of environment from linux
Two utilities, fw_setenv and fw_printenv, as well as a configuration
file, fw_env.config, allow a user to read and modify the u-boot
environment from linux.  This is useful on machines like overo the
have non-volatile storage for a u-boot environment.

Pull in u-boot-fw-utils to builds for overo and provide the custom
fw_env.config needed.  This configuration file should not be shipped
with u-boot itself as u-boot is included using EXTRA_IMAGEDEPENDS
which doesn't get packaged into the rootfs.

Note: The MTD partition for the environment is sized at 256kB for
      spacing although we only use 128kB for the actual environment.

Signed-off-by: Ash Charles <ashcharles@gmail.com>