alsastate: Remove dmix config and clean up state
Since ALSA 1.0.9, digital mixing is enabled by default for soundcards that do
not support hardware mixing. Hence we do not have to provide a config for dmix.

Also, in asound.state, the 2nd half of the asound.state is identical to the 1st
half. This file provides audio codec specific ALSA settings. I believe the file
contains duplicates by error.

Signed-off-by: Ash Charles <ashcharles@gmail.com>