alsastate: Enable mic-in for Gallop43
Unlike the Gumstix Tobi board which uses line-in, Gallop43 uses mic-in for audio
input. This commit unmutes mic-in for a more desirable out-of-box behavior on
Gallop43.

Other than this mixer setting adjustment, no further changes seem to be required.
The audio-routing information specific for twl4030 is completely omitted in
Overo's DTS. Hence it is assumed that all connections are available for use.

Signed-off-by: Ash Charles <ashcharles@gmail.com>