uim: Add TI usermode initialization manager
The UIM tool supports the TI Shared Transport driver from userspace. It
is needed to make bluetooth (and any other UART-based) functions work on
Wilink8 modules.
The original version of this recipe is taken from the meta-ti and
meta-arago layers.  It has been updated and tweaked to support
systemd-based OSes.

Signed-off-by: Ash Charles <ashcharles@gmail.com>