bluez: Add support for bluetooth-via-uart
This reverts commit f9717ff which temporarily disable support for
bluetooth adapters that are wired up using a UART.  Such bluetooth
devices, specifically bluetooth provided on wl18xx and w3cbw0003
connected to ttyO1 on the Gumstix COMs, should be detected and start
automatically on boot---no "hciattach" needed.  This relies on systemd
features but won't break non-systemd systems---they'll just need to
call 'hciattach' some other way.

Note: it is not clear that UIM is needed for the wl18xx chips.  Disable
      this service from starting automatically.

Signed-off-by: Ash Charles <ashcharles@gmail.com>