Enable the generic cpufreq driver for Pepper and build performance and userspace governors.
Fix frequency scaling bug in Pepper which was
caused by incorrect values in device tree for
VDD_CORE and VDD_MPU.

Note: As of now, userspace governor does not seem
to be working on Pepper. Any attempt to change the
frequency using userspace governor results in
irreversibly setting the lowest frequency which
can only be undone by a reboot.

Signed-off-by: Arun Bharadwaj <arun@gumstix.com>