[U-Boot] Expect ext4 as the rootfs type
Gumstix is migrating to EXT4 filesystem type from EXT3 for all 3
platforms: Overo, DuoVero and Pepper. The newly introduced flashing tool
"wic" should be used to generate a compatible image.

Here, migration is handled in two separate patches. One for Overo
and Pepper, which the U-Boot environment variables are set through their
own configuration file. And another for DuoVero, which relies on the
OMAP4's common header file, and is less likely to be upstreamed.

Signed-off-by: Adam YH Lee <adam.yh.lee@gmail.com>