conf: set IMAGE_BOOT_FILES for machines
In order to use the 'wic' tool to create dd-able sd card images,
machines should specify the ${IMAGE_BOOT_FILES}: a list of files that
should be copied from the ${DEPLOY_DIR_IMAGE} to the boot partition. As
discussed [1], a machine, not the u-boot recipe, should specify the
files needed to boot the machine.

[1] http://lists.openembedded.org/pipermail/openembedded-core/2014-October/098458.html

Signed-off-by: Ash Charles <ashcharles@gmail.com>