bluez: Allow systemd to activate bluetooth
By default, the systemd bluetooth.service is enabled so that, if a
bluetooth device is detected, the special bluetooth.target causes the
bluetooth.service to start---this means bluetooth devices get managed
appropriately by bluetoothd.

Without this patch, bluetooth.service is disabled (bbappend from
meta-openembedded) and, for Gumstix boards, (appropriately) gets
started directly in these cases:
 * if a W2CBW003 or Wilink8 wireless module is detected, the systemd
   bluetooth-ttyO1-csr.service or bluetooth-ttyO1-csr.service trigger
   bluetooth.service, or
 * if the booting image includes a utility such as network manager that
   requests the org.bluez service via dbus and, as such, triggers dbus
   activation of the bluetooth.service.
But, if the user hotplugs a bluetooth device or the board uses a
W2CBW0015 wireless module, systemd triggers the bluetooth.target but
bluetooth.service never gets started.  This has meant users have had to
manually "hciconfig hci0 up" or "systemctl <enable|start> bluetooth".

Note: this doesn't always force bluetooth to be active---it'll be
active only when a bluetooth device is detected or a userspace
application requests bluetooth services.

Signed-off-by: Ash Charles <ashcharles@gmail.com>