Make bluez adaptable to bluez 5
gnome-bluetooth and obex depends on Bluez 4. Remove this
dependency in an effort to upgrade to Bluez 5.

These fixes are in discussion in the upstream mailing list.
This is just a work-around for fido and can be safely
reverted after the update.

Use bbappend files in order to support these changes in
the meta-gumstix layer.

Signed-off-by: Arun Bharadwaj <arun@gumstix.com>