Handle Y-Axis inversion in X11 for resistive touch
Resistive touchscreens for Gumstix boards are wired with the bottom
left as the origin but X11 expects the origin to be the upper-left
corner.  Rather than hacking the kernel to invert the y-axis for one
particular driver, handle this in X11.  This will support the variety
of resistive touchscreens used Gumstix boards but we may need to add
some exclusions depending on the particular screen in future.

Signed-off-by: Ash Charles <ashcharles@gmail.com>