linux-gumstix: Enable CDC composite gadget
If a board contains a USB port in peripheral mode (either a USB device
or OTG port), load up the g_cdc module so that it can be used for as a
network or serial connection.  By default, a getty is attached to a
/dev/ttyGS0 port if one exists so users can login over USB---this
commit just restores the behaviour whereby the kernel creates such a
port.

Signed-off-by: Ash Charles <ashcharles@gmail.com>