Updating from Fido to Krogoth
Some files and folders need to be renamed in order to be compatible to Krogoth
Also, updating the defconfig for overo, duovero, and pepper
The last but not least, the patches for Overo, Duovero, and pepper have been applied to kernel version 4.5.
and the newer version of patches will be uploaded soon.

Signed-off-by: Jerry Hung <jerry@gumstix.com>