Fix issues when building the yocto image
Kernel 4.5 has stopped the support for IRQF_DISABLE due to some significant reasons d8bf368
It does not make senses that the interrupts request in the share line can interrupts each other.
As the result, removing IRQF_DISABLE in the pwm.c is necessary.
Initially, the plan is removing whole repository to build the driver; however, Building image will be
failed without this driver.

Signed-off-by: Jerry Hung <jerry@gumstix.com>