Add support for tensorflow (with cuda).
The tensorflow-cuda recipe was largely taken from the tensorflow
recipe in the meta-tensorflow layer.

A recipes to add libcudnn and tensorrt to the native machine was needed to compile
tensorflow with cuda support.

To run tensorflow demos, the python3 libraries kiwisolver and
matplotlib were also required.  Those are not found upstream in the
warrior branch.